Replaced entity framework with a DataAccessLayer project (accesses SQLite directly via sql as EntityFramework wasn't flexible enough and expects you to have full control over the database schema and structure the database in a certain way. We're just reading from an existing schema we have no control over)
Split the database reading into a separate project to keep it clean

Returning matching tags is temporarily disabled while I get it work with the new DataAccessLayer

Data retrieval is now a lot faster due to being more efficient with the SQL and not working around entity framework's stubborness

Repositories now inherit from interfaces to make transitions and testing easier.

Connection strings and client files are now relative to the configured Hydrus Network directory

(breaking change): Project is now only compatible with win64 platforms (replace e_sqlite3.dll with a different one to support a different platform. It has to be in the project root and set to "copy to output" in order for it to be copied into the bin folder.

Added a test Hydrus Network database to the test project to make sure the test data is as expected